extract common code between the ref instance filter and the instance sharing plugins
